requirements.txt
pyserial
cryptography
websockets
flask>=3.0.0
flask-sock
pyjwt
bcrypt
